mcp/slack/pyproject.toml
[project]
name = "slack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.6.0",
    "openai>=1.76.0",
    "python-dotenv>=1.1.0",
]
